--- app/(personal)/(pages)/family/layout.js
@@ -1,6 +1,0 @@
-export const metadata = {
-  title: "Morgan Ney's Family",
-}
-export default function FamilyLayout({ children }) {
-  return children
-}
