--- app/(personal)/(pages)/music/layout.js
@@ -1,6 +1,0 @@
-export const metadata = {
-  title: "Morgan Ney's Playlist",
-}
-export default function MusicLayout({ children }) {
-  return children
-}
