+++ app/(personal)/globals.css
@@ -35,21 +35,28 @@ a:visited {
 a:hover {
   text-decoration: none;
 }
-h1 {
-  margin-top: 0;
-  font-size: 28px;
-  word-break: break-all;
-}
 p {
   margin: 0 0 16px 0;
   line-height: 1.25;
 }
-nav {
-  margin: 0 0 24px 0;
-}
 main,
 footer {
   min-width: 0;
+}
+main > h1,
+main > section:first-child h1 {
+  margin: 0;
+  font-size: 28px;
+  word-break: break-all;
+}
+main > h1 {
+  margin: 0 0 24px 0;
+}
+main > section:first-child {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin: 0 0 24px 0;
 }
 main + footer address {
   display: flex;
